refactor(D-168 B): decompose probability_tutor_brain into 5 contiguous mixins + brain manifest
- probability_brain/ package: escape_hatch (D-124/125/160/162/165) +
  cognitive_verification (D-127 L3/L5) + cognitive_response (D-127 L4) +
  socratic_narrative (D-128/129/130) + turn_engine (D-158/159) — verbatim
  contiguous line-range moves (base order == original line order, so
  first-occurrence anchors in the composed source are preserved structurally)
- probability_tutor_brain.py becomes a ~55-line composition root; the
  OrchestratorClient(ProbabilityTutorBrain, ...) MRO contract is untouched
- brain_sources.py manifest (BRAIN_SOURCE_FILES + read_brain_source, stdlib)
- check_pedagogical_os: _read_brain() reads the brain manifest textually;
  all 10 _read(BRAIN_PATH) sites + the parity gate switched; new consistency
  check: BRAIN_SOURCE_FILES must be a subset of TUTOR_SOURCE_FILES
- check_skills_doctrine: _read_brain_source() via importlib (bypasses the
  pydantic-laden skills __init__ — degraded-sandbox property preserved)
- test_iss128 _client_src() -> read_tutor_source() (kills the last manual
  pre-D-164 concatenation)
- 4 brain workflows extend their glob with probability_brain/*.py
- TUTOR_SOURCE_FILES += the 5 mixins

Brain: 2,441 -> 55-line root + 5 mixins (largest 933). All 1,541
tests/services pass; both gates green; MRO membership verified live.
