+++ .github/workflows/iss-052-bac-display-streaming.yml
@@ -216,6 +216,7 @@ jobs:
           _tutor = [
               _base / 'orchestrator_client.py',
               pathlib.Path('app/services/skills/probability_tutor_brain.py'),
+              *sorted(pathlib.Path('app/services/skills/probability_brain').glob('*.py')),
               *sorted((_base / 'orchestrator').glob('*.py')),
           ]
           tutor_src = '\n'.join(p.read_text() for p in _tutor)
