+++ .github/workflows/microservices-d045-user-routing.yml
@@ -95,6 +95,7 @@ jobs:
           _files = [
               _base / 'orchestrator_client.py',
               pathlib.Path('app/services/skills/probability_tutor_brain.py'),
+              *sorted(pathlib.Path('app/services/skills/probability_brain').glob('*.py')),
               *sorted((_base / 'orchestrator').glob('*.py')),
           ]
           src = '\n'.join(p.read_text() for p in _files)
