+++ app/services/skills/probability_brain/__init__.py
@@ -1,0 +1,1 @@
+"""حزمة عقل الاحتمالات المُفكَّك (D-168) — mixins مُركَّبة في `ProbabilityTutorBrain`."""
