+++ tests/services/test_iss128_question_speech_act.py
@@ -93,12 +93,11 @@ def _other_combo() -> SimpleNamespace:
 
 
 def _client_src() -> str:
-    # D-163: عقل الاحتمالات استُخرج للوحدة المستقلة — نضمّ الاثنين ليصمد أي pin.
-    return (_REPO_ROOT / "app" / "infrastructure" / "clients" / "orchestrator_client.py").read_text(
-        encoding="utf-8"
-    ) + (_REPO_ROOT / "app" / "services" / "skills" / "probability_tutor_brain.py").read_text(
-        encoding="utf-8"
-    )
+    # D-168: المصدر المُركَّب عبر manifest D-164 (يشمل client + brain + كل الـ mixins)
+    # — يقتل آخر concatenation يدوي من عصر ما قبل المانيفست.
+    from app.infrastructure.clients.orchestrator.tutor_sources import read_tutor_source
+
+    return read_tutor_source()
 
 
 def _port_src() -> str:
